+++ site/components/Components.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ShallowComponent from "robe-react-commons/lib/components/ShallowComponent";
 import Renderer from "./Renderer";
-import { Grid, Col, Collapse, ListGroup, ListGroupItem, InputGroup, Nav, NavItem } from "react-bootstrap";
+import {Grid, Col, Collapse, ListGroup, ListGroupItem, InputGroup, Nav, NavItem} from "react-bootstrap";
 import ComponentList from "./ComponentList";
 import Progress from "progress/Progress";
 import TextInput from "inputs/TextInput";
@@ -11,7 +11,7 @@ import "./style.css"
 
 export default class Components extends ShallowComponent {
 
-    constructor(props: Object) {
+    constructor(props:Object) {
         super(props);
         this.state = {
             componentSelection: window.location.hash.substring(1) === "Components" ? "Components/complex/DataGrid" : window.location.hash.substring(1),
@@ -20,7 +20,7 @@ export default class Components extends ShallowComponent {
         };
     }
 
-    render(): Object {
+    render():Object {
         let componentDetail;
         let componentMenu = [];
         let selectedGroup = window.location.hash.split("/")[1] || this.state.selectedGroup;
@@ -36,7 +36,7 @@ export default class Components extends ShallowComponent {
                         key={`#${item.header}`}
                         onClick={this.__onComponenListClick}
                         active={active}
-                        >
+                    >
                         {item.header}
                     </ListGroupItem>);
                 if (active) {
@@ -48,7 +48,7 @@ export default class Components extends ShallowComponent {
                             json={item.json}
                             sample={item.sample}
                             code={item.code}
-                            />);
+                        />);
                 }
             }
         }
@@ -62,10 +62,10 @@ export default class Components extends ShallowComponent {
                         onChange={this.__onFilterChange}
                         value={this.state.filter}
                         placeholder="Search"
-                        />
+                    />
                     <ListGroup className="parent">
                         <ListGroupItem onClick={this.__onGroupChange.bind(undefined, "complex")}>
-                            <FaIcon code="fa-cubes" fixed={false} />&nbsp;
+                            <FaIcon code="fa-cubes" fixed={false}/>&nbsp;
                             Complex
                         </ListGroupItem>
                         <Collapse in={selectedGroup === "complex"}>
@@ -74,7 +74,7 @@ export default class Components extends ShallowComponent {
                             </div>
                         </Collapse>
                         <ListGroupItem onClick={this.__onGroupChange.bind(undefined, "inputs")}>
-                            <FaIcon code="fa-terminal" fixed={false} />&nbsp;
+                            <FaIcon code="fa-terminal" fixed={false}/>&nbsp;
                             Inputs
                         </ListGroupItem>
                         <Collapse in={selectedGroup === "inputs"}>
@@ -83,7 +83,7 @@ export default class Components extends ShallowComponent {
                             </div>
                         </Collapse>
                         <ListGroupItem onClick={this.__onGroupChange.bind(undefined, "charts")}>
-                            <FaIcon code="fa-line-chart" fixed={false} />&nbsp;
+                            <FaIcon code="fa-line-chart" fixed={false}/>&nbsp;
                             Charts
                         </ListGroupItem>
                         <Collapse in={selectedGroup === "charts"}>
@@ -101,23 +101,23 @@ export default class Components extends ShallowComponent {
         );
     }
 
-    __onGroupChange = (selectedKey: string) => {
+    __onGroupChange = (selectedKey:string) => {
         let selectedComponent;
         switch (selectedKey) {
             case "complex":
-                {
-                    selectedComponent = "DataGrid";
-                }
+            {
+                selectedComponent = "DataGrid";
+            }
                 break;
             case "inputs":
-                {
-                    selectedComponent = "TextInput";
-                }
+            {
+                selectedComponent = "TextInput";
+            }
                 break;
             case "charts":
-                {
-                    selectedComponent = "AreaChart";
-                }
+            {
+                selectedComponent = "AreaChart";
+            }
                 break;
             default:
         }
@@ -128,13 +128,13 @@ export default class Components extends ShallowComponent {
         });
     };
 
-    __onFilterChange = (e: Object) => {
+    __onFilterChange = (e:Object) => {
         this.setState({
             filter: e.target.value
         });
     };
 
-    __onComponenListClick = (e: Object) => {
+    __onComponenListClick = (e:Object) => {
         this.setState({
             componentSelection: `Components/${this.state.selectedGroup}/${e.target.text}`
         });