+++ site/components/samples/charts/LineChartSample.jsx
@@ -1,31 +1,31 @@
 import React from "react";
-import { ShallowComponent } from "robe-react-commons";
-import LineChart from "robe-react-ui/lib/chart/LineChart";
+import {ShallowComponent} from "robe-react-commons";
+import LineChart from "robe-react-ui/lib/charts/LineChart";
 
 let data = [
-    { name: "Page A", uv: 4000, pv: 2400, amt: 2400, fill: "#8884d8" },
-    { name: "Page B", uv: 3000, pv: 1398, amt: 2210, fill: "#83a6ed" },
-    { name: "Page C", uv: 2000, pv: 9800, amt: 2290, fill: "#8dd1e1" },
-    { name: "Page D", uv: 2780, pv: 3908, amt: 2000, fill: "#82ca9d" },
-    { name: "Page E", uv: 1890, pv: 4800, amt: 2181, fill: "#a4de6c" },
-    { name: "Page F", uv: 2390, pv: 3800, amt: 2500, fill: "#d0ed57" },
-    { name: "Page G", uv: 3490, pv: 4300, amt: 2100, fill: "#ffc658" },
+    {name: "A", pb: 4000, pv: 2400, pr: 2400},
+    {name: "B", pb: 3000, pv: 1398, pr: 2210},
+    {name: "C", pb: 2000, pv: 9800, pr: 2290},
+    {name: "D", pb: 2780, pv: 3908, pr: 2000},
+    {name: "E", pb: 1890, pv: 4800, pr: 2181},
+    {name: "F", pb: 2390, pv: 3800, pr: 2500},
+    {name: "G", pb: 3490, pv: 4300, pr: 2100}
 ];
+let lines = [
+    {dataKey: "pb", name: "Public", unit: "piece"},
+    {dataKey: "pr", name: "Protected", unit: "piece"},
+    {dataKey: "pv", name: "Private", unit: "piece"}
+];
+
+
 export default class LineChartSample extends ShallowComponent {
 
-    render(): Object {
+    render():Object {
         return (
             <div>
                 <div className="form-group">
-                    <LineChart
-                        name="line-chart"
-                        propsOfChart={{ width: 600, height: 400, data: data }}
-                        propsOfChildrens={[{ dataKey: "uv", stroke: "red", fill: "red" }, { dataKey: "pv", stroke: "blue", fill: "blue" }]}
-                        propsOfXAxis={{ dataKey: "name" }}
-                        propsOfYAxis propsOfToolTip propsOfCartesianGrid propsOfLegend
-                    />
+                    <LineChart data={data} width={400} height={250} lines={lines}/>
                 </div>
-                <a rel="noopener noreferrer" target="_blank" href="http://recharts.org/api/#LineChart">Read More About LineChart</a>
             </div>
         );
     }